p4i integration Integrate 256587 from 4.3 to main: Fixes:    Compilation on Leopard RevBy:    Trenton
[git-p4: depot-paths = "//depot/qt/main/": change = 256588]

+++ src/gui/kernel/qwidget_mac.cpp
@@ -1554,7 +1554,7 @@ void QWidgetPrivate::createWindow_sys()
 
     if (qt_mac_is_macsheet(q)){
         SetThemeWindowBackground(qt_mac_window_for(q), kThemeBrushSheetBackgroundTransparent, true);
-        float alpha = 0;
+        CGFloat alpha = 0;
         GetWindowAlpha(qt_mac_window_for(q), &alpha);
         if (alpha == 1){
             // For some reason the 'SetThemeWindowBackground' does not seem
